feat: separate registration from plan selection and payment
Backend:
- Make plan_id optional in BusinessRegisterRequest (sometimes instead of required)
- BusinessService::register skips subscription creation when no plan_id
- BusinessController only loads subscription.plan when plan_id was sent

Frontend:
- Add /onboarding route for plan selection + onboarding fee payment
- OnboardingPage: PlanCards to select plan, then subscribe + pay in one flow
- AccountQueries: useRegisterBusiness redirects to /onboarding when
  no subscription or onboarding_fee unpaid (instead of /register/payment)
- Add useSubscribe mutation to AccountQueries
- Register SUBSCRIBE and ONBOARDING route constants

+++ database/migrations/2026_07_21_000003_create_payments_table.php
@@ -11,7 +11,7 @@ return new class extends Migration
         Schema::create('billing_payments', function (Blueprint $table) {
             $table->id();
             $table->unsignedBigInteger('business_id');
-            $table->unsignedBigInteger('subscription_id')->nullable();
+            $table->unsignedInteger('subscription_id')->nullable();
             $table->decimal('amount', 14, 2);
             $table->string('currency', 3)->default('UGX');
             $table->string('method', 50)->nullable();
